% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Design.R
\name{addOffset4ModelFrame}
\alias{addOffset4ModelFrame}
\title{Internal function to Calculate Offset for Data}
\usage{
addOffset4ModelFrame(Terms, newdata, offset = 0)
}
\arguments{
\item{Terms}{term value}

\item{newdata}{data frame for prediction. Each row of the data
frame   contains values of covariates that are required in the crr
model. If   missing, the original data set that was used to
develop the crr model   will be used for prediction.}

\item{offset}{offset value}
}
\value{
newdata
}
\description{
Internal function to Calculate Offset for Data
}
\details{
Internally calculate offset for a data
}
