% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QHScrnomo-package.R
\docType{data}
\name{cr.dat}
\alias{cr.dat}
\title{competing risks data}
\format{
A data frame with 1000 observations on the following 5 variables.
  \describe{ \item{list("ftime")}{follow up time}
  \item{list("fstatus")}{event status}
   \item{list("age")}{patient age}
  \item{list("gender")}{Gender}
   \item{list("smoker")}{is a smoker ?}
}
}
\description{
Data set cr.dat
}
\details{
Another artificial dataset
}
\examples{

data(cr.dat)
head(cr.dat)
}
\keyword{datasets}
