% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomo2.crr.R
\name{nomo2.crr}
\alias{nomo2.crr}
\title{Estimate Cumulative Incidence Rate}
\usage{
nomo2.crr(x, f.crr, time)
}
\arguments{
\item{x}{a vector of sum of linear predictors for each subject.}

\item{f.crr}{a saved model fitted by function
\code{\link[QHScrnomo]{crr.fit}}}

\item{time}{expected evaluation time}
}
\value{
a vector with each element being the predicted cumulative incidence
  rate at the expected time.
}
\description{
Calculate Estimated Cumulative Incidence Rate
Calculate predicted cumulative incidence rate based on a competing risks
regression model.
}
\details{
This function is usually used to transform regular failure probabilities to
competing risks adjusted probabilities, when a nomogram of competing risks
regression model is constructed started from a regular survival model. It
is not often called externally.
}
\note{
internal function
}
\seealso{
\code{\link[QHScrnomo]{pred2.crr}}
  \code{\link[QHScrnomo]{crr.fit}} \code{\link[cmprsk]{crr}}
}
\author{
Michael W. Kattan, Ph.D. and Changhong Yu.\cr Department of
  Quantitative Health Sciences, Cleveland Clinic
}
\keyword{survival}
\keyword{utilities}
