% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred3.crr.R
\name{pred3.crr}
\alias{pred3.crr}
\title{Predict cumulative incidence used internally}
\usage{
pred3.crr(z, cov1, cov2, time, lps = FALSE)
}
\arguments{
\item{z}{the fitter crr model}

\item{cov1}{covarite matrix 1}

\item{cov2}{covariate matrix 2}

\item{time}{time point at which the prediction will make}

\item{lps}{logical flag. if the liner predictor be generated.}
}
\value{
A list. See \code{\link[cmprsk]{crr}} for details.
}
\description{
Predict cumulative incidence used internally
an internal function.
}
\details{
Internally used only.
}
\note{
an internal function called by \code{\link{crr.fit}}.
}
\author{
changhong
}
