% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.R
\name{cindex}
\alias{cindex}
\title{Concordance index calculation}
\usage{
cindex(
  prob,
  fstatus,
  ftime,
  type = "crr",
  failcode = 1,
  cencode = 0,
  tol = 1e-20
)
}
\arguments{
\item{prob}{predicted risk of failure event, either probability or
risk score}

\item{fstatus}{failure(event) variable}

\item{ftime}{follow-up time variable for survival or competing risks
predictions}

\item{type}{type of regression models corresponding to different type
of outcomes.
'logis' is the default value for binary outcome, 'surv' for ordinary
survival outcome
and 'crr' for competing risks outcome.}

\item{failcode}{coding for failure(event). 1 is the default value.}

\item{cencode}{coding for censoring. 0 is the defaul}

\item{tol}{error tolerance. the default value is 1e-20.}
}
\value{
a vector of returned values.
\item{N}{the total number of observations in the input data}
\item{n}{the nonmissing number of observations that was used f
or calculation}
\item{usable}{the total number of usable pairs.}
\item{oncordant}{the number of concordant pairs}
\item{cindex}{the concordance index that equal to the number of
concordant pairs
divided by the total number of usable pairs.}
}
\description{
Calculate concordance index
}
\details{
to calculate the discrimination metric, concordance index for binary,
time-to event and competing risks outcomes
}
\examples{
seed <- 5555

# simulated data to test the function
n <- 20000
ftime <- rexp(n)
fstatus <- sample(0:2,n,replace=TRUE)
cov <- matrix(runif(3*n),nrow=n)
dimnames(cov)[[2]] <- c('x1','x2','x3')
dat <- data.frame(ftime,fstatus,cov)
require(Design)
print(z <- cph(Surv(ftime,fstatus == 1) ~ x1 + x2 + x3,data = dat))

# summary(z)
z.p <- predict(z)

#debug(cindex)
cindex(prob=z.p,fstatus=fstatus,type="logis")
cindex(prob=z.p,fstatus=fstatus,ftime=ftime,type="surv")
cindex(prob=z.p,fstatus=fstatus,ftime=ftime,type="crr")

}
\author{
Changhong Yu, Michael Kattan, Brian Wells, Amy Nowacki.
}
\keyword{regression}
\keyword{semiparametric}
