% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred2.crr.R
\name{pred2.crr}
\alias{pred2.crr}
\title{Predict Cumulative Incidence Rate}
\usage{
pred2.crr(f.crr, lp, time)
}
\arguments{
\item{f.crr}{a saved competing risks regression model created by function
\code{\link{crr.fit}}}

\item{lp}{a scalar being the sum of linear predictors for a single subject.}

\item{time}{expected time point, at which cumulative incidence rate will be
assessed.}
}
\value{
Return the predicted cumulative incidence rate.
}
\description{
Predict Cumulative Incidence Rate
}
\details{
Calculate the predicted cumulative incidence rate based on a saved
competing risks regression model.
The cumulative incidence is adjusted for other competing causes
rather than the event of interest.
}
\examples{

data(prostate.dat)
library(Hmisc,TRUE)
library(rms,TRUE)
dd <- datadist(prostate.dat)
options( datadist = "dd")
f.cph <- cph(formula = Surv(TIME_EVENT, EVENT_DOD == 1 ) ~ rcs(AGE,3) +
    CLIN_STG + rcs(PSA, 3),
    data = prostate.dat, x = TRUE, y = TRUE, surv = TRUE)
f.crr <- crr.fit(fit = f.cph, cencode = 0, failcode = 1)

# Estimate cumulative incidence rate by 6 year
QHScrnomo:::pred2.crr(prostate.crr, lp = 0.8, time = 60)

}
\seealso{
\code{\link[cmprsk]{predict.crr}} \code{\link[QHScrnomo]{crr.fit}}
}
\author{
Michael Kattan, Ph.D, Changhong Yu \cr Department of Quantitative
  Health Sciences\cr Cleveland Clinic\cr
}
\keyword{survival}
\keyword{utilities}
