% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Design.R
\name{predictDesign}
\alias{predictDesign}
\title{Internal function to Calculate Predicted Competing Risks Probability.}
\usage{
predictDesign(
  fit,
  newdata = NULL,
  type = c("lp", "x", "data.frame", "terms", "adjto", "adjto.data.frame",
    "model.frame"),
  se.fit = FALSE,
  conf.int = FALSE,
  conf.type = c("mean", "individual"),
  incl.non.slopes = NULL,
  non.slopes = NULL,
  kint = 1,
  na.action = na.keep,
  expand.na = TRUE,
  center.terms = TRUE,
  ...
)
}
\arguments{
\item{fit}{a saved crr model fit crated by function
\code{crr.fit}}

\item{newdata}{data frame for prediction. Each row of the data
frame   contains values of covariates that are required in the crr
model. If   missing, the original data set that was used to
develop the crr model   will be used for prediction.}
}
\value{
A vector with the length equal to the number of rows in the data
  frame, which was used to make prediction. Each element corresponds to a
  predicted failure probability at the expected time point.
}
\description{
Internal function to Calculate Predicted Competing Risks Probability.
}
\details{
Calculate predicted probabilities for a competing risks regression model
}
